pyproject.toml
[tool.poetry]
name = "pattern-matching-nester"
version = "0.1.0"
description = ""
authors = ["Undine Stelter <undine.stelter@mailbox.tu-dresden.de>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
svgpathtools = "^1.7.0"
svgwrite = "^1.4.3"
svg-path = "^6.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
